Upgrade PoW 1024
