--- neurons/deployment_layer/model_1e6fcdaea58741e7248b631718dda90398a17b294480beb12ce8232e27ca3bff/metadata.json
@@ -1,13 +1,0 @@
-{
-    "name": "SN2 PoW 1024 Circom",
-    "description": "Generates batched proof of weights based on miner evaluation data from SN2.",
-    "author": "Inference Labs",
-    "version": "0.0.6",
-    "proof_system": "CIRCOM",
-    "netuid": "2",
-    "type": "proof_of_weights",
-    "external_files": {
-        "circuit.zkey": "https://storage.omron.ai/1e6fcdaea58741e7248b631718dda90398a17b294480beb12ce8232e27ca3bff/circuit.zkey"
-    },
-    "benchmark_choice_weight": 0.2
-}
