--- neurons/deployment_layer/model_1e6fcdaea58741e7248b631718dda90398a17b294480beb12ce8232e27ca3bff/settings.json
@@ -1,57 +1,0 @@
-{
-    "scaling": 100000000,
-    "public_inputs": {
-        "order": [
-            "new_score",
-            "block_number_out",
-            "miner_uid_out",
-            "validator_uid_out",
-            "RATE_OF_DECAY",
-            "RATE_OF_RECOVERY",
-            "FLATTENING_COEFFICIENT",
-            "PROOF_SIZE_THRESHOLD",
-            "PROOF_SIZE_WEIGHT",
-            "RESPONSE_TIME_WEIGHT",
-            "MAXIMUM_RESPONSE_TIME_DECIMAL",
-            "COMPETITION_WEIGHT",
-            "maximum_score",
-            "previous_score",
-            "verified",
-            "proof_size",
-            "response_time",
-            "competition",
-            "maximum_response_time",
-            "minimum_response_time",
-            "block_number",
-            "validator_uid",
-            "miner_uid",
-            "scaling"
-        ],
-        "sizes": {
-            "new_score": 1024,
-            "block_number_out": 1024,
-            "miner_uid_out": 1024,
-            "validator_uid_out": 1024,
-            "RATE_OF_DECAY": 1,
-            "RATE_OF_RECOVERY": 1,
-            "FLATTENING_COEFFICIENT": 1,
-            "PROOF_SIZE_THRESHOLD": 1,
-            "PROOF_SIZE_WEIGHT": 1,
-            "RESPONSE_TIME_WEIGHT": 1,
-            "COMPETITION_WEIGHT": 1,
-            "MAXIMUM_RESPONSE_TIME_DECIMAL": 1,
-            "maximum_score": 1024,
-            "previous_score": 1024,
-            "verified": 1024,
-            "proof_size": 1024,
-            "response_time": 1024,
-            "competition": 1024,
-            "maximum_response_time": 1024,
-            "minimum_response_time": 1024,
-            "block_number": 1024,
-            "validator_uid": 1024,
-            "miner_uid": 1024,
-            "scaling": 1
-        }
-    }
-}
